paper_reproductions/[2-09]_Framelet_Tubular/requirements.txt
# 小波框架管状结构分割依赖
# Python >= 3.8

# 核心数值计算
numpy>=1.20.0
scipy>=1.7.0

# 小波变换
PyWavelets>=1.1.0

# 图像处理
scikit-image>=0.18.0
Pillow>=8.0.0

# 可视化
matplotlib>=3.3.0

# 进度条
tqdm>=4.60.0

# 可选: 医学图像格式支持
# SimpleITK>=2.0.0
# pydicom>=2.0.0

# 可选: GPU加速
# cupy-cuda11x>=9.0.0
